true	FacultyController.java	-/Modified	Commit	src\main\java\YogaStudio\Controller\FacultyController.java true	UserController.java	-/Modified	Commit	src\main\java\YogaStudio\Controller\UserController.java true	WaiverDAO.java	-/Modified	Commit	src\main\java\YogaStudio\dao\generic\WaiverDAO.java true	WaiverService.java	-/Modified	Commit	src\main\java\YogaStudio\service\WaiverService.java true	rejectWaiverPopUp.jsp	-/Modified	Commit	src\main\webapp\WEB-INF\views\waiver\rejectWaiverPopUp.jsp true	viewWaiversByFA.jsp	-/Modified	Commit	src\main\webapp\WEB-INF\views\waiver\viewWaiversByFA.jsp true	waiverDetails.jsp	-/Added	Commit	src\main\webapp\WEB-INF\views\waiver\waiverDetails.jsp

+++ src/main/java/YogaStudio/dao/generic/WaiverDAO.java
@@ -39,7 +39,7 @@ public class WaiverDAO {
         sf.getCurrentSession().persist(waiver);
     }
 
-    public WaiverEntity get(int id) {
+    public WaiverEntity get(Long id) {
         return (WaiverEntity) sf.getCurrentSession().load(WaiverEntity.class, id);
     }
 
@@ -48,7 +48,7 @@ public class WaiverDAO {
         sf.getCurrentSession().update(waiver);
     }
 
-    public void delete(int waiverId) {
+    public void delete(Long waiverId) {
         WaiverEntity waiver = get(waiverId);
         sf.getCurrentSession().delete(waiver);
     }
